# api/openapi.yaml
openapi: 3.0.3
info:
  title: livestream-viewer API
  version: 0.1.0
  description: |
    REST API for livestream-viewer. The management UI is a consumer of this API.
    Interactive docs are served at `/docs`.

    Request bodies are capped at 1 MiB, except `POST /api/v1/uploads` (multipart
    up to 512 MiB) and `PUT /api/v1/system/youtube/cookies` (multipart up to
    256 KiB). Exceeding a cap yields 400 or 413.
servers:
  - url: /
    description: This instance
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      security: []
      summary: Authenticate and receive a JWT
      description: |
        Attempts are rate limited per client address: five in quick
        succession, then one more every twenty seconds.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: "#/components/schemas/User" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/change-password:
    post:
      tags: [auth]
      summary: Change the current user's password
      description: Required after first login with the seeded admin, and after an administrator creates an account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string, format: password }
                new_password: { type: string, format: password, minLength: 8 }
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          description: |
            Malformed body, wrong current password, or a new password shorter
            than 8 characters. Note this is a 400, not a 403 — the 403
            `password_change_required` is raised by *other* endpoints while a
            change is pending.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/users:
    get:
      tags: [users]
      summary: List management users
      description: Admin only.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: "#/components/schemas/User" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [users]
      summary: Create a management user
      description: Admin only. The new user must change their password on first login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password, minLength: 8 }
                role: { type: string, enum: [admin, user] }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Username already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/users/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    patch:
      tags: [users]
      summary: Update a user's role
      description: Admin only. An administrator cannot change their own role.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, user] }
      responses:
        "204":
          description: Role updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [users]
      summary: Delete a management user
      description: Admin only. An administrator cannot delete themselves.
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/system/youtube:
    get:
      tags: [system]
      summary: YouTube session and PO token provider status (never the cookie jar)
      description: |
        `pot.mode` is `managed`, `external`, or `off` (`auto` is resolved at
        startup). `pot.running` means `/ping` succeeded. That is not a
        guarantee of playback. `last_error_code` may be `youtube_bot_check`
        or `youtube_auth` from live ingest.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/YouTubeStatus" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/system/youtube/cookies:
    put:
      tags: [system]
      summary: Upload a Netscape cookies.txt for yt-dlp
      description: |
        Used when YouTube still treats this host as a bot after the PO token
        provider is running, and for private, members-only, or age-restricted
        videos. Cookie values are never returned.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/YouTubeStatus" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: Cookie file exceeds 256 KiB
    delete:
      tags: [system]
      summary: Remove the YouTube cookie jar
      responses:
        "200":
          description: Removed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/YouTubeStatus" }
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/system/youtube/po-token:
    put:
      tags: [system]
      summary: Set or clear the optional YouTube PO token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                po_token:
                  type: string
                  description: Empty string clears the stored token.
      responses:
        "200":
          description: Stored
          content:
            application/json:
              schema: { $ref: "#/components/schemas/YouTubeStatus" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/config:
    get:
      tags: [config]
      summary: Runtime configuration
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeConfig" }
    patch:
      tags: [config]
      summary: Update runtime configuration
      description: |
        Admin only. A sparse patch: only the keys present in the body are
        applied, and `id` is ignored. Saving reloads the display strategy.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RuntimeConfigPatch" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RuntimeConfig" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/sources:
    get:
      tags: [sources]
      summary: List sources
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources:
                    type: array
                    items: { $ref: "#/components/schemas/Source" }
    post:
      tags: [sources]
      summary: Create a source
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceWrite" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/sources/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [sources]
      summary: Get a source
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [sources]
      summary: Update a source
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceWrite" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [sources]
      summary: Delete a source
      description: Refused if the source is referenced by a screen; the error details list those screens.
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/SourceInUse"
  /api/v1/sources/{id}/probe:
    parameters:
      - $ref: "#/components/parameters/ID"
    post:
      tags: [sources]
      summary: Re-probe codec, resolution, hardware decode
      description: |
        Opens the stream and runs a decode test, so it costs about as much as a
        live tile. At most two probes run at once; further calls are rejected
        rather than queued.

        On failure, `probe.message` carries the diagnostic with any URL
        credentials redacted, truncated to 512 characters.
      responses:
        "200":
          description: Probe result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Source" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/v1/sources/{id}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [sources]
      summary: Cached still for the UI
      description: |
        Returns a JPEG under `{data.dir}/thumbnails/{id}.jpg`. Probing writes
        the still, so a source that has never probed successfully returns 404.

        Because browsers cannot set an `Authorization` header on `<img src>`,
        this endpoint also accepts the JWT in the `access_token` query
        parameter. Prefer the header wherever the caller controls it.
      parameters:
        - name: access_token
          in: query
          required: false
          description: JWT, as an alternative to the `Authorization` header.
          schema: { type: string }
      responses:
        "200":
          description: JPEG
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/uploads:
    get:
      tags: [uploads]
      summary: List uploaded files
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  uploads:
                    type: array
                    items: { $ref: "#/components/schemas/Upload" }
    post:
      tags: [uploads]
      summary: Upload a media file
      description: |
        Maximum upload size is 512 MiB. Allowed MIME types are common video formats
        (for example video/mp4, video/webm, video/quicktime).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Upload" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: File too large
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /api/v1/uploads/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    delete:
      tags: [uploads]
      summary: Remove an uploaded file
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/UploadInUse"
  /api/v1/screens:
    get:
      tags: [screens]
      summary: List screens
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  screens:
                    type: array
                    items: { $ref: "#/components/schemas/Screen" }
    post:
      tags: [screens]
      summary: Create a screen
      description: Admin only. Grid screens default to the `full` layout when `layout` is omitted.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScreenWrite" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Screen" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/v1/screens/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [screens]
      summary: Get a screen including tiles and playlist items
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Screen" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [screens]
      summary: Update a screen
      description: |
        Admin only. A sparse patch: `tiles` and `items` are replaced wholesale
        when present and left untouched when omitted. Changing `layout` remaps
        existing tiles onto the new cell count.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ScreenWrite" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Screen" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [screens]
      summary: Delete a screen
      responses:
        "204":
          description: Deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/ScreenInUse"
  /api/v1/tour:
    get:
      tags: [tour]
      summary: The screen tour
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tour" }
    put:
      tags: [tour]
      summary: Replace the screen tour
      description: |
        Admin only. Full replacement of the tour document. The `entries` array
        is required (use `[]` to clear a disabled tour); omitting it is a 400.
        Entry `position` is assigned from array order, so any value sent is
        ignored.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [entries]
              properties:
                enabled: { type: boolean }
                loop: { type: boolean }
                entries:
                  type: array
                  items: { $ref: "#/components/schemas/TourEntry" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Tour" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/v1/display/next:
    post:
      tags: [display]
      summary: Advance the tour
      responses:
        "204":
          description: Advanced
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/display/previous:
    post:
      tags: [display]
      summary: Step back
      responses:
        "204":
          description: Stepped
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/display/goto/{screenId}:
    post:
      tags: [display]
      summary: Jump to a screen and pin it
      parameters:
        - name: screenId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Pinned
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/display/pause:
    post:
      tags: [display]
      summary: Hold the tour
      responses:
        "204":
          description: Paused
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/display/resume:
    post:
      tags: [display]
      summary: Release the tour
      responses:
        "204":
          description: Resumed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/preview/stream:
    get:
      tags: [preview]
      summary: MJPEG of the composited output
      description: |
        At most four viewers are served at once; each holds a connection for as
        long as it stays subscribed.
      responses:
        "200":
          description: multipart MJPEG
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/EngineNotRunning"
  /api/v1/preview/frame:
    get:
      tags: [preview]
      summary: Single JPEG snapshot
      responses:
        "200":
          description: JPEG
          content:
            image/jpeg:
              schema: { type: string, format: binary }
        "503":
          $ref: "#/components/responses/PreviewUnavailable"
  /api/v1/events:
    get:
      tags: [display]
      summary: Server-Sent Events for display and decoder state
      description: |
        Streams `event: display.state` with a JSON `DisplayState` payload whenever
        scheduler state changes (plus comment keepalives). EventSource clients that
        cannot set an Authorization header may pass `?access_token=<jwt>`.

        At most sixteen clients are served at once.
      parameters:
        - name: access_token
          in: query
          required: false
          schema: { type: string }
          description: JWT for EventSource clients (alternative to Authorization header)
      responses:
        "200":
          description: SSE stream (`text/event-stream`)
          content:
            text/event-stream:
              schema: { type: string }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/ServerError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: integer }
  responses:
...
    BadRequest:
      description: Malformed body, invalid path id, or failed validation
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Authenticated but not an administrator, or a password change is pending
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ServerError:
      description: Database or I/O failure
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    SourceInUse:
      description: |
        Source is referenced by one or more screens. `details.screens` lists
        them so the UI can offer to clear the references.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SourceConflict" }
...
    RateLimited:
      description: |
        A resource guard rejected the request. `Retry-After` carries the
        suggested delay in seconds. Error codes are `rate_limited` for login
        attempts, `probe_busy` when the concurrent probe limit is reached, and
        `too_many_clients` on the streaming endpoints.
      headers:
        Retry-After:
          schema: { type: integer }
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }
        details: {}
...
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: string, enum: [admin, user] }
        must_change_password: { type: boolean }
...
    RuntimeConfig:
      type: object
      properties:
        id: { type: integer }
        output_width: { type: integer }
        output_height: { type: integer }
        default_dwell_ms: { type: integer }
        default_transition: { $ref: "#/components/schemas/Transition" }
        allow_software_fallback: { type: boolean }
        max_hw_decoders: { type: integer }
        reconnect_backoff_ms: { type: integer }
        preview_fps: { type: integer }
        preview_width: { type: integer }
        placeholder_color: { type: string }
        gutter_px: { type: integer }
...
    YouTubeStatus:
      type: object
      description: Process-level YouTube session and provider health. Cookie values are never returned.
      properties:
        configured: { type: boolean }
        youtube_hosts: { type: integer }
        updated_at: { type: string, format: date-time, nullable: true }
        po_token: { type: boolean }
        last_error_code:
          type: string
          description: Live ingest classifier, for example `youtube_bot_check` or `youtube_auth`.
        last_error: { type: string }
        pot:
          type: object
          description: BgUtils PO token provider health. Not a guarantee of playback.
          properties:
            mode:
              type: string
              enum: [managed, external, off]
            running: { type: boolean }
            base_url: { type: string }
            version: { type: string }
            last_error: { type: string }